backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3
sentence-transformers==2.7.0
transformers==4.40.0
torch==2.1.1
pinecone
python-dotenv==1.0.0
tqdm==4.66.1
pydantic==2.5.0
numpy<2.0
slowapi==0.1.9
cachetools==5.3.2
validators==0.22.0
